+++ types.ts
@@ -62,7 +62,6 @@ export type MessageTypesT =
   | 'planner_result'
   | 'planner_followup'
   | 'aimode_search_action'
-  | 'page_qa_endpoint'
 
 export type PromptDataT = {
   prompt: string
